+++ CSVToESLib/Types/Result.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace CSVToESLib.Template
+namespace CSVToESLib.Types
 {
     public class Result<T, U>
     {
